移除ToolBox.ts
